classes: add kernel-fitimage-overlay class
This adds support for a fitImage having DT overlays in it.

It will iterate over space-separated list KERNEL_DEVICETREE_OVERLAYS to add
an image and a configuration per DTBO.

The load address DTB_LOAD for Device Tree has to be specified as well as
load address DTBO_LOAD for Device Tree Overlays.

The configuration per DTBO will be named after the DTBO filename.

Signed-off-by: Quentin Schulz <quentin.schulz@free-electrons.com>